middleware/requirements.txt
Flask==2.3.2
paho-mqtt==1.6.1
influxdb~=5.3.1
python-dotenv==1.0.0
requests~=2.31.0